refactor(Money): Money의 amount를 double로 변경
 - Money에서 연산자 오버라이딩을 하다가 div의 결과는 double인데 Money는 double값을 가질 수 없는 것을 확인했다. 그래서 amount의 자료형을 double로 변경하고 long을 받을 수 있는 부생성자를 추가했다

+++ src/main/kotlin/domain/WinStatistics.kt
@@ -25,7 +25,7 @@ class WinStatistics(winningNumbers: WinningNumbers, purchasedLottoBundle: LottoB
     }
 
     private fun calculateEarningRate(spentMoney: Money): Double {
-        return getTotalIncome() / spentMoney
+        return (getTotalIncome() / spentMoney).toDouble()
     }
 
     override fun equals(other: Any?): Boolean {
